feat(okms): use <BillingInformationsTileStandard />
ref: #MANAGER-16644
