+++ packages/manager/apps/key-management-service/package.json
@@ -18,6 +18,7 @@
     "test:coverage": "vitest run --coverage"
   },
   "dependencies": {
+    "@ovh-ux/billing-informations": "^1.2.21",
     "@ovh-ux/logs-to-customer": "^1.6.1",
     "@ovh-ux/manager-config": "^8.5.1",
     "@ovh-ux/manager-core-api": "^0.17.1",
