--- packages/manager/apps/key-management-service/src/components/layout-helpers/Dashboard/GeneralInformationsTiles/BillingInformationsTile.tsx
@@ -1,181 +1,0 @@
-import React, { useEffect, useContext } from 'react';
-import {
-  ActionMenu,
-  ActionMenuItem,
-  DashboardTile,
-  DashboardTileBlockItem,
-  LinkType,
-  Links,
-} from '@ovh-ux/manager-react-components';
-import {
-  useFeatureAvailability,
-  ServiceDetails,
-} from '@ovh-ux/manager-module-common-api';
-import {
-  ODS_ICON_NAME,
-  ODS_TEXT_PRESET,
-  ODS_BADGE_COLOR,
-  ODS_BUTTON_VARIANT,
-} from '@ovhcloud/ods-components';
-import { Outlet, useNavigate } from 'react-router-dom';
-import { OdsText, OdsBadge } from '@ovhcloud/ods-components/react';
-import { useTranslation } from 'react-i18next';
-import {
-  ButtonType,
-  PageLocation,
-  ShellContext,
-  useOvhTracking,
-} from '@ovh-ux/manager-react-shell-client';
-import { OkmsServiceState } from '../okmsServiceState/OkmsServiceState.component';
-import { TileValueDate } from '@/components/dashboard/tile-value-date/tileValueDate.component';
-import { dateFormat } from './BillingInformationsTile.constants';
-import { KMS_ROUTES_URIS } from '@/routes/routes.constants';
-import { KMS_FEATURES } from '@/utils/feature-availability/feature-availability.constants';
-
-type BillingInformationsTileProps = {
-  okmsService: ServiceDetails;
-};
-
-const BillingInformationsTile = ({
-  okmsService,
-}: BillingInformationsTileProps) => {
-  const { t } = useTranslation('key-management-service/dashboard');
-  const navigate = useNavigate();
-  const [contactUrl, setContactUrl] = React.useState('#');
-  const { trackClick } = useOvhTracking();
-  const { data: availability } = useFeatureAvailability([
-    KMS_FEATURES.DISPLAY_CONTACTS,
-  ]);
-  const {
-    shell: { navigation },
-  } = useContext(ShellContext);
-
-  // Generate the contact URL for managing contacts
-  useEffect(() => {
-    navigation
-      .getURL('dedicated', '#/contacts/services', {})
-      .then((url) => setContactUrl(url as string));
-  }, [navigation]);
-
-  const menuItems: ActionMenuItem[] = [
-    {
-      id: 1,
-      label: t('key_management_service_dashboard_action_billing_terminate'),
-      onClick: () => {
-        trackClick({
-          location: PageLocation.page,
-          buttonType: ButtonType.link,
-          actionType: 'navigation',
-          actions: ['delete_kms'],
-        });
-        navigate(KMS_ROUTES_URIS.kmsTerminate);
-      },
-    },
-  ];
-
-  const tileItems: DashboardTileBlockItem[] = [
-    {
-      id: 'creationDate',
-      label: t('key_management_service_dashboard_field_label_creation_date'),
-      value: (
-        <TileValueDate
-          value={okmsService?.billing.lifecycle.current.creationDate}
-          options={dateFormat}
-        />
-      ),
-    },
-    {
-      id: 'state',
-      label: t('key_management_service_dashboard_field_label_state'),
-      value: (
-        <span>
-          <OkmsServiceState state={okmsService?.resource?.state} />
-        </span>
-      ),
-    },
-    {
-      id: 'nextBillingDate',
-      label: t(
-        'key_management_service_dashboard_field_label_next_billing_date',
-      ),
-      value: (
-        <div className="flex flex-row justify-between items-center gap-2">
-          <TileValueDate
-            value={okmsService?.billing.nextBillingDate}
-            options={dateFormat}
-          />
-
-          <div className="flex flex-row align-center gap-4">
-            <ActionMenu
-              id="nextBillingDateMenu"
-              items={menuItems}
-              isCompact
-              icon={ODS_ICON_NAME.ellipsisVertical}
-              variant={ODS_BUTTON_VARIANT.ghost}
-            />
-          </div>
-        </div>
-      ),
-    },
-    {
-      id: 'engagement',
-      label: t('key_management_service_dashboard_field_label_engagement'),
-      value: (
-        <span>
-          <OdsBadge
-            color={ODS_BADGE_COLOR.critical}
-            label={
-              okmsService?.billing.engagement?.endRule?.strategy ??
-              t('key_management_service_dashboard_field_label_engagement_none')
-            }
-          />
-        </span>
-      ),
-    },
-  ];
-
-  if (availability?.[KMS_FEATURES.DISPLAY_CONTACTS])
-    tileItems.push({
-      id: 'contacts',
-      label: t('key_management_service_dashboard_field_label_contacts'),
-      value: (
-        <>
-          <div className="mb-3">
-            {okmsService?.customer.contacts.map((contact) => (
-              <OdsText
-                key={contact.customerCode + contact.type}
-                preset={ODS_TEXT_PRESET.span}
-                className="block mb-1"
-              >{`${contact.customerCode} ${t(
-                `key_management_service_dashboard_contact_type_${contact.type}`,
-              )}`}</OdsText>
-            ))}
-          </div>
-          <Links
-            type={LinkType.next}
-            href={contactUrl}
-            onClickReturn={() =>
-              trackClick({
-                location: PageLocation.page,
-                buttonType: ButtonType.externalLink,
-                actionType: 'navigation',
-                actions: ['contact_support'],
-              })
-            }
-            label={t(
-              'key_management_service_dashboard_field_label_manage_contacts',
-            )}
-          />
-        </>
-      ),
-    });
-
-  return (
-    <>
-      <DashboardTile title={t('billing_informations')} items={tileItems} />
-      <Outlet />
-    </>
-  );
-};
-
-export default BillingInformationsTile;
